resolve issue; move test to proper place
